+++ NEWS.md
@@ -2,8 +2,12 @@
 
 * Rebasing for the purposes of Bioconductor version 2.2.0
 
-## UNDERLYING CHANGES
+## NEW FEATURES
 * Added support for BCRs for loading ParseBio sequences.
+* Added `quietBCRgenes()` and `quietTCRgenes()` from Ibex and Trex and `quietVDJgenes()` as a convenience that runs both. The functions filter out known TCR and/or BCR gene signatures.
+
+## UNDERLYING CHANGES
+* Added `Seurat` to the `Suggests` field in the DESCRIPTION file.
 
 # scRepertoire VERSION 2.0.8